pyproject.toml
[tool.poetry]
name = "python3-poetry-pyenv"
version = "0.1.0"
description = ""
authors = ["Leonardo Franco <leonardomf96@gmail.com>"]
readme = "README.md"
packages = [{include = "python3_poetry_pyenv"}]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.2.3"
flask-sqlalchemy = "^3.0.3"
flask-migrate = "^4.0.4"
flask-restful = "^0.3.9"
flask-apispec = "^0.11.4"
flask-marshmallow = "^0.15.0"
marshmallow-sqlalchemy = "^0.29.0"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.6"
flask-cors = "^3.0.10"
openai = "^0.27.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
env_files = ["envs/.env.test"]